Tested and fixed with -Wconversion
git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/tools/vislcg3/trunk@5054 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/Grammar.h
@@ -39,7 +39,7 @@ namespace CG3 {
 		bool has_dep;
 		bool has_encl_final;
 		bool is_binary;
-		uint32_t grammar_size;
+		size_t grammar_size;
 		UChar mapping_prefix;
 		uint32_t lines;
 		mutable double total_time;
